refactor(health): rename /ping to /health; unify handler, response, and test names
