+++ internal/health/handler.go
@@ -1,4 +1,4 @@
-package ping
+package health
 
 import (
 	"encoding/json"
@@ -11,18 +11,18 @@ import (
 // @Description Returns "pong" to indicate the server is alive
 // @Tags health
 // @Produce json
-// @Success 200 {object} ping.PingResponse "pong message"
-// @Router /ping [get]
-func NewPingHandler(logger *zap.Logger) http.HandlerFunc {
+// @Success 200 {object} health.HealthgResponse "pong message"
+// @Router /health [get]
+func NewHealthHandler(logger *zap.Logger) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		logger.Info("responding to /ping")
+		logger.Info("responding to /health")
 
 		w.Header().Set("Content-Type", "application/json")
 		w.WriteHeader(http.StatusOK)
 
-		resp := PingResponse{Message: "pong"}
+		resp := HealthResponse{Message: "pong"}
 		if err := json.NewEncoder(w).Encode(resp); err != nil {
-			logger.Error("failed to write ping response", zap.Error(err))
+			logger.Error("failed to write health response", zap.Error(err))
 		}
 	}
 }
