+++ internal/health/response.go
@@ -1,0 +1,7 @@
+package health
+
+// HealthResponse represents a simple health check response.
+// Typically used for /health endpoint.
+type HealthResponse struct {
+	Message string `json:"message" example:"pong"`
+}
